Clean up filenames and style in python editor
